+++ changelog.md
@@ -11,6 +11,7 @@
 - Better validation for Woodland Warriors urls
 - Clearer wording on /schedule set. "Poll" is reworded to "Suggest"
 - Suggesting a time pings the players
+- /lfg's host only actions now give a hint to press "Join"
 
 ### Bug Fixes
 - js was breaking due to auto filling box score to game form